Consolidated metadata related classes in the database.metadata package, including the Options and Recogniser classes
