+++ src/main/java/nz/co/gregs/dbvolution/databases/metadata/ForeignKeyRecognisor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nz.co.gregs.dbvolution.generation;
+package nz.co.gregs.dbvolution.databases.metadata;
 
 /**
  * A Helper class to capture naming conventions of databases.
