+++ src/main/java/nz/co/gregs/dbvolution/databases/metadata/Options.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package nz.co.gregs.dbvolution.generation;
+package nz.co.gregs.dbvolution.databases.metadata;
 
 import java.util.Arrays;
 import nz.co.gregs.dbvolution.databases.DBDatabase;
