+++ src/main/java/nz/co/gregs/dbvolution/databases/metadata/PrimaryKeyRecognisor.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nz.co.gregs.dbvolution.generation;
+package nz.co.gregs.dbvolution.databases.metadata;
 
 /**
  * A Helper class to capture naming conventions of databases.
