+++ src/main/java/nz/co/gregs/dbvolution/generation/DataRepo.java
@@ -30,6 +30,7 @@
  */
 package nz.co.gregs.dbvolution.generation;
 
+import nz.co.gregs.dbvolution.databases.metadata.Options;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.ArrayList;
